Code refactorization
