+++ bitblas/gpu/matmul_mma_dequantize.py
@@ -9,7 +9,7 @@ from contextlib import suppress
 from tvm import tir, DataType
 from tvm.target import Target
 
-from ..ops.common import TransformKind
+from bitblas.base.operator_common import TransformKind
 from ..base.roller.hint import Hint, IntrinInfo
 from ..base.roller.rasterization import NoRasterization
 from ..base import analysis
