+++ bitblas/ops/general_matmul_splitk.py
@@ -4,7 +4,7 @@ from tvm.target import Target
 import operator
 from functools import reduce
 from typing import Any, Optional, Union
-from .common import TransformKind
+from bitblas.base.operator_common import TransformKind
 from .impl.matmul_splitk_impl import select_implementation as consistent_implementation
 from .impl.matmul_dequantize_splitk_impl import select_implementation as weight_dequantize_implementation
 from dataclasses import dataclass
