+++ bitblas/ops/impl/matmul_dequantize_impl.py
@@ -4,7 +4,7 @@
 from bitblas import tvm
 from tvm import te, DataType
 from tvm.tir import IndexMap
-from bitblas.ops.common import TransformKind
+from bitblas.base.operator_common import TransformKind
 from bitblas.gpu.matmul_analysis import get_propagate_map
 from bitblas.quantization import (
     _tir_packed_int_to_int_convert,
